add: last message preview.
